+++ fortran/script/generate_fit_tau.py
@@ -96,7 +96,7 @@ def generate_fit_tau_function(ndim, input_type, output_type):
   status_c = {c_function}(obj%tau_smpl_ptr, obj%backend_ptr, SPIR_ORDER_COLUMN_MAJOR, &
     ndim_c, c_loc(input_dims_c), target_dim_c, c_loc(arr), c_loc(res))
   IF (status_c /= 0) THEN
-    CALL errore('fit_tau_{func_suffix}_{ndim}d', 'Error fitting on tau sampling points', status_c)
+    CALL errore('fit_tau_{func_suffix}_{ndim}d', 'Error fitting on tau sampling points', INT(status_c))
   ENDIF
 END SUBROUTINE"""
